base: Timestamp type on arm_64 is 64bit

+++ repos/base/include/spec/arm_64/trace/timestamp.h
@@ -19,7 +19,7 @@
 
 namespace Genode { namespace Trace {
 
-	typedef uint32_t Timestamp;
+	typedef uint64_t Timestamp;
 
 	inline Timestamp timestamp()
 	{
